Passado o Id e RA, de integer para LONG e também adicionado o findbyid

+++ src/main/java/com/br/SAM_FullStack/SAM_FullStack/service/AlunoService.java
@@ -20,13 +20,14 @@ public class AlunoService {
     private EmailService emailService;
 
 
-    public Aluno findById(Integer id) {
+    public Aluno findById(Long id) {
         return alunoRepository.findById(id).orElseThrow(() ->
                 new RuntimeException("Aluno não encontrado com ID: " + id));
     }
 
     public Aluno findByRa(Integer ra){
-        return alunoRepository.findByRa(ra).orElseThrow(() -> new RuntimeException("Aluno não encontrado com RA: " + ra));
+        return alunoRepository.findByRa(ra).orElseThrow(() ->
+                new RuntimeException("Aluno não encontrado com RA: " + ra));
     }
 
     public List<Aluno> findAll(){
@@ -57,7 +58,7 @@ public class AlunoService {
         }
     }
 
-    public Aluno update(Integer id, Aluno alunoUpdate){
+    public Aluno update(Long id, Aluno alunoUpdate){
 
         Aluno alunoExistente = findById(id);
         alunoExistente.setNome(alunoUpdate.getNome());
@@ -69,7 +70,7 @@ public class AlunoService {
         return alunoRepository.save(alunoExistente);
     }
 
-    public void delete(Integer id){
+    public void delete(Long id){
         Aluno aluno = findById(id);
         alunoRepository.delete(aluno);
     }
